Add files via upload
Created a fully responsive Netflix Clone using only HTML and CSS, showcasing exceptional attention to detail, design accuracy, and responsive layout. This project is built solely with the power of visualisation and well-curated resources, reflecting my commitment to crafting a seamless and professional user experience. A testament to minimalist coding, this front-end project exemplifies the art of pure HTML and CSS development.